+++ packages/react-paypal-js/src/v6/components/Braintree/BraintreePayPalProvider.test.tsx
@@ -4,7 +4,7 @@ import { act, render, waitFor } from "@testing-library/react";
 
 import { expectCurrentErrorValue } from "../../hooks/useErrorTestUtil";
 import { BraintreePayPalProvider } from "./BraintreePayPalProvider";
-import { useBraintreePayPal } from "../../hooks/useBraintreePayPal";
+import { useBraintreePayPal } from "../../hooks/Braintree/useBraintreePayPal";
 import { INSTANCE_LOADING_STATE } from "../../types/ProviderEnums";
 
 import type { BraintreeV6Namespace } from "../../types";
